Improve robustness of textual HMM serialization
The current code for textual serialization of HMMs suffers from two
problems:
1. HmmWriter and OpdfWriter use DecimalFormat for converting numbers to
strings. Since no locale is specified, the default locale of the JVM is
used, which breaks the unit tests on systems that do not use '.' as the
decimal separator. Furthermore, the specified format string limits the
precision to six decimal places, so that smaller probabilities are
rounded down to zero. In consequence, a HMM may be corrupted by a round-
trip through HmmWriter and HmmReader.
2. In other places, e.g. ObservationRealWriter, numbers are converted to
string implicitly by string concatenation, which for double variables is
equivalent to calling Double.toString. For any double value x, this
method is guaranteed to produce a textual representation that is
independent of the current locale and preserves the exact value of x,
which is what we want. However, it will use scientific notation for very
large values and values close to zero. HmmReader and related classes use
StreamTokenizer, which does not support scientific notation, and most
likely never will (see
http://bugs.java.com/bugdatabase/view_bug.do?bug_id=4079180).

Solve these problems by consistently using implicit double to string
conversion everywhere, and delegating the parsing of decimal numbers to
Double.parseDouble. Remove some redundant code in
ObservationSequenceReader along the way.