Simplify the implementation of KMeansLearner and fix a bug in KMeansCalculator
In the constructor of KMeansLearner, the elements of the current cluster
are iterated over by index, with the index of the currently examined
element being stored in variable j. If that element is moved to another
cluster, the indices of the subsequent elements shift down by one, but j
is incremented unconditionally at the next iteration, causing the next
element to be skipped. Fix this by using an iterator.

Since we do not require random access to the elements in a cluster, we
can replace the ArrayList that serves as backing storage with a
LinkedList, which performs the operations we care about in O(1):
insertion at the end and deletion at the current iterator position,
which are O(1) (amortized) and O(n), respectively, for ArrayList.