Allow users to provide custom implementations of Math.log/exp
A major bottleneck of parameter fitting in log-space are the repeated
calls to Math.log/exp. An optimized implementation of these methods,
such as the one found in the "jafama" library, can speed up the Baum-
Welch algorithm up to 5x, depending on the JRE version.

However, jafama is somewhat inconvenient to use: In order to enable the
optimized version of Math.log, a system property has to be set at JVM
startup.

The mechanism proposed in this patch lets the caller decide which
implementation to use. For example, the default (JRE Math) can be
overridden as follows:

import net.jafama.FastMath;
...
LogSpace.INST = new LogSpace() {
	public double expImpl(double x) { return FastMath.exp(x); }
	public double logImpl(double x) { return FastMath.log(x); }
};

In my benchmarks, the additional layer of indirection did not have a
measurable impact on performance.